chore: move breadcrumbs to own folder
